apply sorting to filters, too
