+++ src/data/bugs/index.tsx
@@ -56,3 +56,40 @@ export interface Bug {
     notes?: string;
     logs?: { label: string; url: string }[];
 }
+
+// builds #s for ptr are lower than live as of 6/19, ugly fix to sort builds including the full patch num at the top as they are technically newer for ptr version
+// genuinely hate this impl and will remove it once build #s go above but probably need a diff permanent fix for later
+const parseBuild = (entry: string): { patch: number[]; build: number } => {
+    const parts = entry.split(".");
+    const build = parseInt(parts[parts.length - 1]);
+    const patch = parts.slice(0, -1).map(Number);
+    return { patch, build };
+};
+
+const compareBuilds = (a: string, b: string): number => {
+    const { patch: patchA, build: buildA } = parseBuild(a);
+    const { patch: patchB, build: buildB } = parseBuild(b);
+    const len = Math.max(patchA.length, patchB.length);
+    for (let i = 0; i < len; i++) {
+        const diff = (patchA[i] ?? 0) - (patchB[i] ?? 0);
+        if (diff !== 0) return diff;
+    }
+    return buildA - buildB;
+};
+
+const getLatestBuildEntry = (buildsTested: string[]): string | undefined =>
+    buildsTested.reduce<string | undefined>((latest, entry) =>
+        !latest || compareBuilds(entry, latest) > 0 ? entry : latest, undefined);
+
+export const getLatestBuild = (buildsTested: string[]): number => {
+    const latest = getLatestBuildEntry(buildsTested);
+    return latest ? parseBuild(latest).build : 0;
+};
+
+export const getBuildSortValue = (buildsTested: string[]): number => {
+    const latest = getLatestBuildEntry(buildsTested);
+    if (!latest) return 0;
+    const { patch, build } = parseBuild(latest);
+    const patchScore = patch.reduce((acc, n) => acc * 1000 + n, 0);
+    return patchScore * 1e9 + build;
+};
